// src/main/interfaces/ITaskData.ts
export default interface ITaskData {
  groupName: any;
  checkoutType: string;
  custom: boolean;
  password: string;
  email: string;
  category: string;
  checkoutDelay: number;
  captchaBypass?: boolean;
  color: string;
  id: number;
  monitorDelay: number;
  monitorType?: string;
  monitorInput: Array<string> | string;
  productColor: string;
  profile: string;
  promoCode: string;
  proxies: string;
  site: string;
  mode: string;
  siteName: string;
  siteType: string;
  sizes: Array<string>;
  status: string;
  retryDelay?: number;
  use3ds?: boolean;
  preRelease: boolean;
  quantity?: string;
  headlessMode?: boolean;
  restockMode?: boolean;
  urlMode?: boolean;
  requireLogin?: boolean;
  useAccountPool?: boolean;
  useShippingRate?: boolean;
  useStoreCredit?: boolean;
  accountPool?: string[];
  shippingRate?: string;
  type: 'script' | 'browser';
}